Adding StatisticsManager
