+++ src/OrleansRuntime/Scheduler/PoliciedScheduler/SchedulingStrategies/BoundaryBasedEDFSchedulingStrategy.cs
@@ -74,8 +74,8 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler.SchedulingStrategies
             {
                 // Leaves a footprint
                 controllerContext.ActivationSeen.Add(schedulingContext.Activation.Address); 
-                if (!workitemManager.UpstreamOpSet.Contains(invokeWorkItem.SourceActivation.Grain))
-                    workitemManager.UpstreamOpSet.Add(invokeWorkItem.SourceActivation.Grain);
+                if (!workitemManager.StatManager.UpstreamOpSet.Contains(invokeWorkItem.SourceActivation.Grain))
+                    workitemManager.StatManager.UpstreamOpSet.Add(invokeWorkItem.SourceActivation.Grain);
             }
             
         }
@@ -97,7 +97,7 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler.SchedulingStrategies
             }
 
             var wim = wig.WorkItemManager as BoundaryBasedEDFWorkItemManager;
-            wim.GetDownstreamContext(downstreamActivation, downstreamContext);
+            wim.StatManager.GetDownstreamContext(downstreamActivation, downstreamContext);
         }
 
         public WorkItemGroup CreateWorkItemGroup(IOrleansTaskScheduler ots, ISchedulingContext context)
@@ -124,7 +124,7 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler.SchedulingStrategies
             if (addressToWIG.ContainsKey(sendingActivationAddress))
             {
                 var wim = addressToWIG[sendingActivationAddress].WorkItemManager as BoundaryBasedEDFWorkItemManager;
-                return wim.CheckForStatsUpdate(upstream);
+                return wim.StatManager.CheckForStatsUpdate(upstream);
             }
             return null;
         }
@@ -161,25 +161,21 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler.SchedulingStrategies
         private FixedSizedQueue<long> queuingDelays;
         private readonly Stopwatch stopwatch;
 #endif
-
-        internal ConcurrentBag<GrainId> UpstreamOpSet { get; set; } // upstream Ops, populated during initialization
-        private ConcurrentDictionary<ActivationAddress, long> DownstreamOpToCost { get; set; } // downstream Ops, populated while downstream message flows back
-
-        private ConcurrentDictionary<GrainId, Tuple<Dictionary<string, long>, long>> StatsUpdatesCollection { get; set; }
-
-        public BoundaryBasedEDFSchedulingStrategy Strategy { get; set; }
         internal long DataflowSLA { get; set; }
+
+        private BoundaryBasedEDFSchedulingStrategy Strategy { get; set; }
+
+        public StatisticsManager StatManager { get; set; }
+
         public bool WindowedGrain { get; set; }
         public long WindowSize { get; set; }
 
         public BoundaryBasedEDFWorkItemManager(ISchedulingStrategy strategy, WorkItemGroup wig)
         {
             Strategy = (BoundaryBasedEDFSchedulingStrategy)strategy;
+            StatManager = new StatisticsManager(wig);
             workItems = new SortedDictionary<long, Queue<Task>>();
             timestampsToDeadlines = new SortedDictionary<long, long>();
-            UpstreamOpSet = new ConcurrentBag<GrainId>();
-            DownstreamOpToCost = new ConcurrentDictionary<ActivationAddress, long>();
-            StatsUpdatesCollection = new ConcurrentDictionary<GrainId, Tuple<Dictionary<string, long>, long>>();
             DataflowSLA = SchedulerConstants.DEFAULT_DATAFLOW_SLA;
             _logger = LogManager.GetLogger(this.GetType().FullName, LoggerType.Runtime);
             workItemGroup = wig;
@@ -221,7 +217,7 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler.SchedulingStrategies
                 {
                     maximumDownStreamPathCost = SchedulerConstants.DEFAULT_WIG_EXECUTION_COST;
 
-                    if (DownstreamOpToCost.Any()) maximumDownStreamPathCost = DownstreamOpToCost.Values.Max();
+                    if (StatManager.DownstreamOpToCost.Any()) maximumDownStreamPathCost = StatManager.DownstreamOpToCost.Values.Max();
 
                     var ownerStats = workItemGroup.WorkItemGroupStats;
                     if (contextObj.SourceActivation != null &&
@@ -347,26 +343,7 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler.SchedulingStrategies
 
         public void UpdateWIGStatistics()
         {
-            if (--statCollectionCounter <= 0)
-            {
-                statCollectionCounter = SchedulerConstants.MEASUREMENT_PERIOD_WORKITEM_COUNT;
-                workItemGroup.CollectStats();
-                var statsToReport = workItemGroup.WorkItemGroupStats;
-                foreach (var address in statsToReport.Keys)
-                {
-                    var statsUpdate = statsToReport[address];
-                    var downstreamCost = DownstreamOpToCost.Values.Any()
-                        ? DownstreamOpToCost.Values.Max()
-                        : SchedulerConstants.DEFAULT_WIG_EXECUTION_COST;
-#if PQ_DEBUG
-                    _logger.Info($"{System.Reflection.MethodBase.GetCurrentMethod().Name} {workItemGroup.Name} -> {address} {string.Join(",", statsUpdate.Select(kv => kv.Key + "->" + kv.Value))} {downstreamCost}");
-#endif
-                    var tup = new Tuple<Dictionary<string, long>, long>(statsUpdate, downstreamCost);
-                    StatsUpdatesCollection.AddOrUpdate(address.Grain, tup, (k, v) => tup);
-
-                }
-                // workItemGroup.LogExecTimeCounters();
-            }
+            StatManager.UpdateWIGStatistics();
         }
 
         public int CountWIGTasks()
@@ -419,28 +396,6 @@ namespace Orleans.Runtime.Scheduler.PoliciedScheduler.SchedulingStrategies
 
             wig.PriorityContext = new PriorityObject(priority, Environment.TickCount);
             dequeuedFlag = true;
-        }
-
-
-        public void GetDownstreamContext(ActivationAddress downstreamActivation, DownstreamContext downstreamContext)
-        {
-            // TODO: FIX LATER
-#if PQ_DEBUG
-                    _logger.Info($"{System.Reflection.MethodBase.GetCurrentMethod().Name} {workItemGroup} <- {downstreamActivation} {downstreamContext}");
-#endif
-            var maxDownstreamCost = downstreamContext.MaximumDownstreamCost +
-                                    downstreamContext.ExecutionCostByTaskType.Count > 0 ? downstreamContext.ExecutionCostByTaskType.Values.Max() : SchedulerConstants.DEFAULT_WIG_EXECUTION_COST;
-            DownstreamOpToCost.AddOrUpdate(downstreamActivation, maxDownstreamCost, (k, v) => maxDownstreamCost);
-        }
-
-        public DownstreamContext CheckForStatsUpdate(GrainId upstream)
-        {
-            Tuple<Dictionary<string, long>, long> tuple;
-            if (StatsUpdatesCollection.TryGetValue(upstream, out tuple))
-            {
-                return new DownstreamContext(tuple.Item1, tuple.Item2);
-            }
-            return null;
         }
 
         public long PeekNextDeadline()
